gravitee-am-certificate/gravitee-am-certificate-api/src/test/java/io/gravitee/am/certificate/api/jwk/JwkNimbusConverterTest.java
/**
 * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *         http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package io.gravitee.am.certificate.api.jwk;

import com.nimbusds.jose.jwk.Curve;
import com.nimbusds.jose.jwk.ECKey;
import com.nimbusds.jose.jwk.JWK;
import com.nimbusds.jose.jwk.RSAKey;
import org.junit.jupiter.api.Test;

import java.security.KeyPair;
import java.security.KeyPairGenerator;
import java.security.interfaces.ECPublicKey;
import java.security.interfaces.RSAPublicKey;
import java.util.Set;

import static org.junit.jupiter.api.Assertions.assertTrue;

class JwkNimbusConverterTest {

    @Test
    void should_return_rsa_nimbus_converter() throws Exception{
        KeyPairGenerator keyPairGenerator = KeyPairGenerator.getInstance("RSA");
        keyPairGenerator.initialize(1024);
        KeyPair keyPair = keyPairGenerator.generateKeyPair();
        RSAPublicKey publicKey = (RSAPublicKey) keyPair.getPublic();

        JWK jwk = new RSAKey.Builder(publicKey).build();

        assertTrue(JwkNimbusConverter.converter(jwk, true, Set.of("sig"), "RSA256") instanceof JwkNimbusRSAConverter);
    }

    @Test
    void should_return_ec_nimbus_converter() throws Exception{
        KeyPairGenerator keyPairGenerator = KeyPairGenerator.getInstance("EC");
        keyPairGenerator.initialize(256);
        KeyPair keyPair = keyPairGenerator.generateKeyPair();
        ECPublicKey publicKey = (ECPublicKey) keyPair.getPublic();

        JWK jwk = new ECKey.Builder(Curve.P_256, publicKey).build();
        assertTrue(JwkNimbusConverter.converter(jwk, true, Set.of("sig"), "EC256") instanceof JwkNimbusECConverter);

    }

}